requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
paho-mqtt==2.1.0
Pygments==2.19.1
pyserial==3.5
rich==14.0.0
